+++ src/app/pages/settings-integration/billing/settings-billing.component.ts
@@ -3,6 +3,7 @@ import { FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import { StatusCodes } from 'http-status-codes';
+import { BillingAccountsTableDataSource } from 'pages/accounts/accounts-table-data-source';
 
 import { CentralServerService } from '../../../services/central-server.service';
 import { ComponentService } from '../../../services/component.service';
@@ -14,7 +15,6 @@ import { HTTPError } from '../../../types/HTTPError';
 import { BillingSetting, BillingSettings, BillingSettingsType, StripeBillingSetting } from '../../../types/Setting';
 import { TenantComponents } from '../../../types/Tenant';
 import { Utils } from '../../../utils/Utils';
-import { BillingAccountsTableDataSource } from './connected-account/settings-billing-accounts-table-data-source';
 
 @Component({
   selector: 'app-settings-billing',
